users/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
psycopg2-binary
bcrypt
python-jose[cryptography]
passlib
python-dotenv
pydantic[email]
python-multipart
